+++ library/Exceptions/CallException.php
@@ -11,6 +11,6 @@
 
 namespace Respect\Validation\Exceptions;
 
-class CallException extends AbstractGroupedException
+class CallException extends GroupedValidationException
 {
 }