+++ library/Exceptions/CallbackException.php
@@ -11,7 +11,7 @@
 
 namespace Respect\Validation\Exceptions;
 
-class CallbackException extends AbstractNestedException
+class CallbackException extends NestedValidationException
 {
     public static $defaultTemplates = [
         self::MODE_DEFAULT => [