+++ CHANGELOG.md
@@ -53,6 +53,8 @@ All notable changes of the Respect\Validation releases are documented in this fi
 - On exceptions, improve `Object` conversion to string (#399)
 - On exceptions, improve conversion of all values by using JSON (#399)
 - On exceptions, nested messages are displayed in a Markdown list (#588)
+- Rename exception class "AbstractGroupedException" to "GroupedValidationException" (#591)
+- Rename exception class "AbstractNestedException" to "NestedValidationException" (#591)
 - Rename rule "Arr" to "ArrayVal"
 - Rename rule "Bool" to "BoolType" (#426)
 - Rename rule "False" to "FalseVal" (#426)